Add new BaseComponent
Make Component and SubComponent inherit from BaseComponent
