+++ src/sst/core/simulation.h
@@ -81,10 +81,6 @@ public:
 
     typedef std::map<SimTime_t, Clock*>   clockMap_t;              /*!< Map of times to clocks */
     typedef std::map<SimTime_t, OneShot*> oneShotMap_t;            /*!< Map of times to OneShots */
-    typedef std::vector<std::string>      statEnableList_t;        /*!< List of Enabled Statistics */
-    typedef std::vector<Params>           statParamsList_t;        /*!< List of Enabled Statistics Parameters */
-    typedef std::map<ComponentId_t, statEnableList_t*> statEnableMap_t;  /*!< Map of Statistics that are requested to be enabled for a component defined in configGraph */
-    typedef std::map<ComponentId_t, statParamsList_t*> statParamsMap_t;  /*!< Map of Statistic Params for a component defined in configGraph */
     // typedef std::map< unsigned int, Sync* > SyncMap_t; /*!< Map of times to Sync Objects */
 
     ~Simulation();
@@ -258,29 +254,6 @@ public:
 		}
     }
 
-    statEnableList_t* getComponentStatisticEnableList(const ComponentId_t &id) const
-    {
-		statEnableMap_t::const_iterator i = statisticEnableMap.find(id);
-		if (i != statisticEnableMap.end()) {
-			return i->second;
-		} else {
-            printf("Simulation::getComponentStatisticEnableList() couldn't find component with id = %lu\n", id);
-            exit(1);
-		}
-        return NULL; 
-    }
-    
-    statParamsList_t* getComponentStatisticParamsList(const ComponentId_t &id) const
-    {
-		statParamsMap_t::const_iterator i = statisticParamsMap.find(id);
-		if (i != statisticParamsMap.end()) {
-			return i->second;
-		} else {
-            printf("Simulation::getComponentStatisticParamsList() couldn't find component with id = %lu\n", id);
-            exit(1);
-		}
-        return NULL; 
-    }
 
 
     /**
@@ -395,8 +368,6 @@ private:
     ThreadSync*      threadSync;
     ComponentInfoMap compInfoMap;
     clockMap_t       clockMap;
-    statEnableMap_t  statisticEnableMap;
-    statParamsMap_t  statisticParamsMap;
     oneShotMap_t     oneShotMap;
     SimTime_t        currentSimCycle;
     SimTime_t        endSimCycle;
