+++ src/sst/core/statapi/componentregisterstat_impl.h
@@ -24,8 +24,8 @@
     bool                            statGood = true;
     bool                            nameFound = false;
     StatisticBase::StatMode_t       statCollectionMode = StatisticBase::STAT_MODE_COUNT;
-    Simulation::statEnableList_t*   statEnableList;
-    Simulation::statParamsList_t*   statParamsList;
+    ComponentInfo::statEnableList_t*   statEnableList;
+    ComponentInfo::statParamsList_t*   statParamsList;
     Output                          out = Simulation::getSimulation()->getSimulationOutput();
     UnitAlgebra                     collectionRate;
     UnitAlgebra                     startAtTime;
@@ -38,7 +38,7 @@
     Statistic<T>*                   statistic = NULL;    
     
     // Check to see if the Statistic is previously registered with the Statistics Engine
-    StatisticBase* prevStat = Simulation::getSimulation()->getStatisticsProcessingEngine()->isStatisticRegisteredWithEngine<T>(getName(), getId(), statName, statSubId);
+    StatisticBase* prevStat = Simulation::getSimulation()->getStatisticsProcessingEngine()->isStatisticRegisteredWithEngine<T>(getName(), my_info->getID(), statName, statSubId);
     if (NULL != prevStat) {
         // Dynamic cast the base stat to the expected type
         return dynamic_cast<Statistic<T>*>(prevStat);
@@ -62,8 +62,8 @@
     // }
 
     // Get Component Statistic Information from the ConfigGraph data
-    statEnableList = Simulation::getSimulation()->getComponentStatisticEnableList(getId());
-    statParamsList = Simulation::getSimulation()->getComponentStatisticParamsList(getId());
+    statEnableList = my_info->getStatEnableList();
+    statParamsList = my_info->getStatParams();
     
     // Check each entry in the StatEnableList (from the ConfigGraph via the 
     // Python File) to see if this Statistic is enabled, then check any of 
@@ -225,7 +225,7 @@
     }
 
     // Register the new Statistic with the Statistic Engine
-    Simulation::getSimulation()->getStatisticsProcessingEngine()->registerStatisticWithEngine<T>(getId(), statistic);
+    Simulation::getSimulation()->getStatisticsProcessingEngine()->registerStatisticWithEngine<T>(my_info->getID(), statistic);
     return statistic;
 //}
