+++ src/sst/core/subcomponent.cc
@@ -10,186 +10,30 @@
 // distribution.
 
 #include <sst_config.h>
-#include "sst/core/component.h"
-#include "sst/core/subcomponent.h"
-#include "sst/core/unitAlgebra.h"
-
-#include <string>
-
-//#include "sst/core/event.h"
-#include "sst/core/factory.h"
-#include "sst/core/link.h"
-#include "sst/core/linkMap.h"
-#include "sst/core/simulation.h"
-#include "sst/core/timeConverter.h"
-#include "sst/core/timeLord.h"
-#include "sst/core/unitAlgebra.h"
-
-using namespace SST::Statistics;
+#include <sst/core/subcomponent.h>
+#include <sst/core/factory.h>
 
 namespace SST {
 
-SubComponent::SubComponent(Component* parent) :
-    parent(parent)
-{
-    type = parent->currentlyLoadingSubComponent;
-}
-
-SubComponent::SubComponent() : 
-    parent(NULL)
-{
-}
-
-SubComponent::~SubComponent() 
-{
-}
-
 bool
-SubComponent::isPortConnected(const std::string &name) const
+SubComponent::doesComponentInfoStatisticExist(const std::string &statisticName)
 {
-    return parent->isPortConnected(name);
+    return Factory::getFactory()->DoesSubComponentInfoStatisticNameExist(my_info->getType(), statisticName);
 }
 
-Link*
-SubComponent::configureLink(std::string name, TimeConverter* time_base, Event::HandlerBase* handler)
+uint8_t SubComponent::getComponentInfoStatisticEnableLevel(const std::string &statisticName)
 {
-    return parent->configureLink(name,time_base,handler);
+    const std::string& type = parent->my_info->getType();
+    /* TODO:  SubComponent, not component stat? */
+    return Factory::getFactory()->GetComponentInfoStatisticEnableLevel(type, statisticName);
 }
 
-Link*
-SubComponent::configureLink(std::string name, std::string time_base, Event::HandlerBase* handler)
+std::string SubComponent::getComponentInfoStatisticUnits(const std::string &statisticName)
 {
-    return configureLink(name,Simulation::getSimulation()->getTimeLord()->getTimeConverter(time_base),handler);
+    const std::string& type = parent->my_info->getType();
+    /* TODO:  SubComponent, not component stat? */
+    return Factory::getFactory()->GetComponentInfoStatisticUnits(type, statisticName);
 }
-
-Link*
-SubComponent::configureLink(std::string name, Event::HandlerBase* handler)
-{
-    return parent->configureLink(name,handler);
-}
-
-Link*
-SubComponent::configureSelfLink( std::string name, TimeConverter* time_base, Event::HandlerBase* handler)
-{
-    return parent->configureSelfLink(name,time_base,handler);
-}
-
-Link*
-SubComponent::configureSelfLink( std::string name, std::string time_base, Event::HandlerBase* handler)
-{
-    return parent->configureSelfLink(name,time_base,handler);
-}
-
-Link*
-SubComponent::configureSelfLink( std::string name, Event::HandlerBase* handler)
-{
-    return parent->configureSelfLink(name,handler);
-}
-
-bool
-SubComponent::doesSubComponentInfoStatisticExist(std::string statisticName)
-{
-    return Factory::getFactory()->DoesSubComponentInfoStatisticNameExist(type, statisticName);
-}
-
-
-TimeConverter* SubComponent::registerClock( std::string freq, Clock::HandlerBase* handler) {
-    return parent->registerClock(freq,handler,false);
-}
-
-TimeConverter* SubComponent::registerClock( const UnitAlgebra& freq, Clock::HandlerBase* handler) {
-    return parent->registerClock(freq,handler,false);
-}
-
-void SubComponent::unregisterClock(TimeConverter *tc, Clock::HandlerBase* handler) {
-    parent->unregisterClock(tc,handler);
-}
-
-Cycle_t SubComponent::reregisterClock( TimeConverter* freq, Clock::HandlerBase* handler) {
-    return parent->reregisterClock(freq,handler);
-}
-
-Cycle_t SubComponent::getNextClockCycle( TimeConverter* freq ) {
-    return Simulation::getSimulation()->getNextClockCycle(freq);
-}
-
-//For now, no OneShot support in SubSubComponent
-#if 0
-TimeConverter* SubComponent::registerOneShot( std::string timeDelay, OneShot::HandlerBase* handler) {
-    return Simulation::getSimulation()->registerOneShot(timeDelay, handler);
-}
-
-TimeConverter* SubComponent::registerOneShot( const UnitAlgebra& timeDelay, OneShot::HandlerBase* handler) {
-    return Simulation::getSimulation()->registerOneShot(timeDelay, handler);
-}
-#endif
-    
-
-TimeConverter*
-SubComponent::getTimeConverter( const std::string& base )
-{
-    return Simulation::getSimulation()->getTimeLord()->getTimeConverter(base);
-}
-    
-TimeConverter*
-SubComponent::getTimeConverter( const UnitAlgebra& base )
-{
-    return Simulation::getSimulation()->getTimeLord()->getTimeConverter(base);
-}
-
-    
-SimTime_t SubComponent::getCurrentSimTime(TimeConverter *tc) const {
-    return tc->convertFromCoreTime(Simulation::getSimulation()->getCurrentSimCycle());
-}
-
-SimTime_t SubComponent::getCurrentSimTime(std::string base) {
-    return getCurrentSimTime(Simulation::getSimulation()->getTimeLord()->getTimeConverter(base));
-
-}
-    
-SimTime_t SubComponent::getCurrentSimTimeNano() const {
-    return getCurrentSimTime(Simulation::getSimulation()->getTimeLord()->getNano());
-}
-
-SimTime_t SubComponent::getCurrentSimTimeMicro() const {
-    return getCurrentSimTime(Simulation::getSimulation()->getTimeLord()->getMicro());
-}
-
-SimTime_t SubComponent::getCurrentSimTimeMilli() const {
-    return getCurrentSimTime(Simulation::getSimulation()->getTimeLord()->getMilli());
-}
-
-Module*
-SubComponent::loadModule(std::string type, Params& params)
-{
-    return parent->loadModule(type,params);
-}
-
-Module*
-SubComponent::loadModuleWithComponent(std::string type, Params& params)
-{
-    return parent->loadModuleWithComponent(type,parent,params);
-
-}
-
-SubComponent*
-SubComponent::loadSubComponent(std::string type, Params& params)
-{
-    return parent->loadSubComponent(type,parent,params);
-}
-
-SharedRegion*
-SubComponent::getLocalSharedRegion(const std::string &key, size_t size)
-{
-    return parent->getLocalSharedRegion(key,size);
-}
-
-SharedRegion*
-SubComponent::getGlobalSharedRegion(const std::string &key, size_t size, SharedRegionMerger *merger)
-{
-    return parent->getGlobalSharedRegion(key, size, merger);
-}
-
 
 
 } // namespace SST
